remove hardcode(string) from report_bottom_sheet
