+++ app/src/main/java/ir/kindnesswall/view/main/charity/charitydetail/CharityDetailActivity.kt
@@ -72,7 +72,7 @@ class CharityDetailActivity : BaseActivity(), CharityViewListener {
 
     private fun sendReport() {
         showMessageDialog(
-            "گزارش تخلف", "ارسال تخلف", false
+            getString(R.string.report_message), getString(R.string.report_send), false
         ) { des ->
             viewModel.getMessageOfReport(
                 CharityReportMessageModel(
@@ -229,39 +229,6 @@ class CharityDetailActivity : BaseActivity(), CharityViewListener {
         StaticContentViewer.show(this, viewModel.charityModel?.website)
     }
 
-    private fun showMessageDialog(
-        title: String?,
-        btnLabel: String?,
-        canceledOnTouchOutside: Boolean,
-        sendRequest: (String) -> Unit
-    ): BottomSheetDialog? = BottomSheetDialog(this).apply {
-        val bind = CustomBottomSheetBinding.bind(
-            layoutInflater.inflate(
-                R.layout.custom_bottom_sheet,
-                null
-            )
-        )
-        setContentView(bind.root)
-        window?.setBackgroundDrawableResource(android.R.color.transparent)
-        setCanceledOnTouchOutside(canceledOnTouchOutside)
-        setCancelable(canceledOnTouchOutside)
-        bind.txtTitle.text = title
-        bind.btnSendRepost.apply {
-            text = btnLabel
-            setOnClickListener {
-                if (bind.txtDescribtion.text.toString().trim().isNullOrEmpty()) {
-                    bind.txtDescribtion.hint = "لطفا گزارش را وارد کنید"
-                } else {
-                    sendRequest(bind.txtDescribtion.text.toString())
-                    dismiss()
-                }
 
-            }
-        }
-        bind.btnCancelRepost.setOnClickListener {
-            dismiss()
-        }
-        if (!isFinishing) if (!isShowing()) show()
-    }
 
 }
