+++ app/src/main/res/layout/report_bottom_sheet.xml
@@ -11,19 +11,17 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:gravity="center"
-        android:text="گزارش تخلف"
+        android:text="@string/report_message"
         android:textColor="@color/red"
         android:textSize="17sp"
         app:layout_constraintTop_toTopOf="parent" />
-
-    <LinearLayout
-        android:id="@+id/linearLayout2"
+    <FrameLayout
+        android:id="@+id/describtion_container"
         android:layout_width="match_parent"
         android:layout_height="270dp"
         android:layout_margin="10dp"
         android:background="@drawable/bg_round_edit_report"
         app:layout_constraintTop_toBottomOf="@id/txtTitle">
-
         <ir.kindnesswall.utils.widgets.LineEditText
             android:id="@+id/txtDescribtion"
             android:layout_width="match_parent"
@@ -38,25 +36,22 @@
             android:maxLines="7"
             android:textAppearance="?android:attr/textAppearanceMedium"
             android:typeface="serif" />
-    </LinearLayout>
-
+    </FrameLayout>
     <TextView
         android:id="@+id/text_world"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_margin="15dp"
+        android:layout_margin="5dp"
         android:text="@string/txt_more_char"
-        app:layout_constraintEnd_toEndOf="@+id/linearLayout2"
-        app:layout_constraintTop_toBottomOf="@+id/linearLayout2" />
-
-
+        app:layout_constraintEnd_toEndOf="@+id/describtion_container"
+        app:layout_constraintTop_toBottomOf="@+id/describtion_container" />
     <TextView
         android:id="@+id/btn_cancelRepost"
         android:layout_width="80dp"
         android:layout_height="40dp"
         android:layout_gravity="center"
         android:gravity="center"
-        android:text="لفو"
+        android:text="@string/cancel_laghve"
         android:layout_marginBottom="10dp"
         android:textColor="@color/colorPrimary"
         android:background="@drawable/bg_round_cancel_report"
@@ -73,7 +68,6 @@
         android:layout_gravity="center"
         android:layout_marginTop="10dp"
         android:gravity="center"
-
         android:layout_marginBottom="10dp"
         android:paddingEnd="10dp"
         android:paddingStart="10dp"
