+++ app/src/main/res/values/strings.xml
@@ -209,5 +209,9 @@
     <string name="all">نمایش برای تمامی کاربر ها</string>
     <string name="enter_report">گزارش را وارد کنید</string>
     <string name="txt_more_char">حداقل ۱۰۰ کاراکتر</string>
+    <string name="report_message">گزارش تخلف</string>
+    <string name="report_send">"ارسال تخلف"</string>
+    <string name="please_enter_repost">""لطفا گزارش را وارد کنید"</string>
+    <string name="cancel_laghve">لفو</string>
 
 </resources>
